Move strip_detector_nms out of trtkit into posekit.artifacts
Review feedback: NMS stripping encodes MMDeploy-detector export patterns
(TopK/Gather/NMS clusters) — model-family surgery, not shared transport.
It now lives beside the OpenMMLab zip acquisition it operates on; trtkit's
onnx_graph keeps only the generic static-batch probe.
