+++ packages/posekit/src/posekit/models/yolox.py
@@ -1,7 +1,7 @@
 """YOLOX-HumanArt person detector on the posekit runtime abstraction.
 
 The ONNX artifact is the OpenMMLab SDK deploy zip (the rtmlib source) with its
-baked-in NMS stripped (:func:`trtkit.onnx_graph.strip_detector_nms`), so ONNX
+baked-in NMS stripped (:func:`posekit.artifacts.strip_detector_nms`), so ONNX
 Runtime and TensorRT share one static-shape graph and thresholding/NMS runs on
 GPU with torchvision.
 """
@@ -13,9 +13,8 @@ from typing import Literal
 import torch
 from jaxtyping import Float, UInt8
 from torch import Tensor
-from trtkit.onnx_graph import strip_detector_nms
 
-from posekit.artifacts import fetch_openmmlab_onnx
+from posekit.artifacts import fetch_openmmlab_onnx, strip_detector_nms
 from posekit.models.base import PersonDetector
 from posekit.ops.letterbox import letterbox_frames
 from posekit.ops.yolox import decode_yolox_head_outputs
